refactor: move createItem to demo.utils
