+++ composables/useDemo.ts
@@ -1,9 +1,8 @@
 import { reactive } from 'vue'
 import itemFixtures from '@/assets/fixtures/items.json'
 import containerFixtures from '@/assets/fixtures/containers.json'
-import { createItem } from '@/utils/item.utils'
 import { Item, ContainerSize, ContainerSlot } from '@/interfaces/inventory'
-import { getRandomItem, generateSlots, depositFirstAvailableSlot } from '@/utils/demo.utils'
+import { getRandomItem, generateSlots, depositFirstAvailableSlot, createItem } from '@/utils/demo.utils'
 import { generateId } from '@/utils/id.utils'
 
 type DemoContainer = {
